Refactor core pipeline into separate modules
